refactor: updated widget locations
